split DbStructureTypes from Types
