+++ src/PostgREST/DbStructure.hs
@@ -41,6 +41,7 @@ import Protolude.Unsafe              (unsafeHead)
 import Text.InterpolatedString.Perl6 (q)
 
 import PostgREST.Private.Common
+import PostgREST.DbStructureTypes
 import PostgREST.Types
 
 getDbStructure :: [Schema] -> [Schema] -> PgVersion -> Bool -> HT.Transaction DbStructure
