+++ src/PostgREST/Parsers.hs
@@ -22,6 +22,7 @@ import Text.ParserCombinators.Parsec hiding (many, (<|>))
 import PostgREST.Error      (ApiRequestError (ParseRequestError))
 import PostgREST.RangeQuery (NonnegRange)
 import PostgREST.Types
+import PostgREST.DbStructureTypes (FieldName)
 import Protolude            hiding (intercalate, option, replace, toS,
                              try)
 import Protolude.Conv       (toS)
