+++ test/Feature/AndOrParamsSpec.hs
@@ -7,7 +7,8 @@ import Test.Hspec
 import Test.Hspec.Wai
 import Test.Hspec.Wai.JSON
 
-import PostgREST.Types (PgVersion, pgVersion112)
+import PostgREST.Types (pgVersion112)
+import PostgREST.DbStructureTypes (PgVersion)
 import Protolude       hiding (get)
 import SpecHelper
