+++ test/Feature/QuerySpec.hs
@@ -8,8 +8,9 @@ import Test.Hspec          hiding (pendingWith)
 import Test.Hspec.Wai
 import Test.Hspec.Wai.JSON
 
-import PostgREST.Types (PgVersion, pgVersion112, pgVersion121,
+import PostgREST.Types (pgVersion112, pgVersion121,
                         pgVersion96)
+import PostgREST.DbStructureTypes (PgVersion)
 import Protolude       hiding (get)
 import SpecHelper
